+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # CHANGE LOG
 
 ## [0.15.1]
+
 - Fix absolute link is prefixed by host issue
 
 ## [0.15.0]
 
+- add features and tests for advanced search
 - add features and tests for item details pages
 - add features and tests for EDD requests
 - add features and tests for item details pages